Create complex_numbers.analogy
Here is a version of an analogy file which uses hdtp to determine the analogy between a real closed field and a one-domesional vector space. This version only finds the obvious analogy between the existing closed fields. The problem here is a lack of backtracking in hdtp - also it is not clear how to explicitly quantify - in this version I have used types and _Z for existential variables - this is not correct but needs sorting.